Two-word Algo
Restructured database to record previous two words and utilize that for
greatly increased accuracy. This required a rebuild of the database
(which also includes a one-word table to use as a fallback should the
two-word fail to find a match). This took 40 hours for the same texts,
hence its inclusion here.