Complete Code Overhaul
Started over in a fresh file (new name is also representative of a
recent move to Python 3.4) and now use a three-word-default algorithm.
The results are impressive.

In this version I also significantly optimized the SQL database creation
to the point where the database (now running on a RAMDisk) can be
generated in less than 10 minutes and the SQL queries are no longer the
bottleneck in this. With the three-word, this release also uses a single
table, building the two- and one-word lists from the same data on use.
The newfound speed has allowed for significant improvements to the text
pre- and post-processing, due to minimal "recompile" times now.

Removing the sample database, as the creation process should be quick to
duplicate now and this will eliminate any possible copyright concerns
(however baseless).