Kinda-sorta Rewrite
Further refined the pre- and post-processing of text for the database.
Stripped out illegal characters and made adjustments to assist in the
respecting of punctuation and case.

Introduced cross-sentence context to improve flow, which required a
re-write of both database building and sentence generation. Sentences
are no longer treated as unique entities, instead pull context from the
end of the previous sentence. Paragraphs are now generated per word and
sentences are metered out during this process by tracking punctuation.