SPEED AND POWER
That's right, even after finding additional texts, I've cut the total
processing time from no database to a 4.6Million record database and 4
"chapter" generated text saved in under 5 minutes.

This is largely from a restructuring of the in-memory databases used to
generate the text. This has resulted in a large memory footprint (peaked
around 3.4Gb in my tests), so it would likely require 64-bit Python to
even run, but the performance gains from this are phenomenal, with no
side-effects as of yet.