License Change
Due to the possibility of this project going public, I've decided to
change the licensing of this from its previous MIT to GPLv3. The license
file in the project has been updated to reflect this and should be
considered the license for all iterations of this project.

My reason for changing this is not to limit usage and expansion of this
project. On the contrary, after much deliberation, I realized that I
wanted people using this in a way that was more open and transparent,
such that this could help expand the project at large. Depending on the
attention this may (or may not) get, I also wanted to ensure that if a
derivative of this project was used or distributed in another
application, that too would be equally open-source.

I've also included the latest version of the program itself, which
corrects a few bugs and added more conditions to the pre- and
post-processing phases of the text.